// src/api-services/models/paginated-response-item-simple-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 规范化接口演示
 * 让 .NET 开发更简单，更通用，更流行。
 *
 * OpenAPI spec version: 1.0.0
 * Contact: monksoul@outlook.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ItemSimpleDto } from "./item-simple-dto"
/**
 *
 * @export
 * @interface PaginatedResponseItemSimpleDto
 */
export interface PaginatedResponseItemSimpleDto {
   /**
    *
    * @type {number}
    * @memberof PaginatedResponseItemSimpleDto
    */
   pageNumber?: number
   /**
    *
    * @type {number}
    * @memberof PaginatedResponseItemSimpleDto
    */
   pageSize?: number
   /**
    *
    * @type {number}
    * @memberof PaginatedResponseItemSimpleDto
    */
   totalNumber?: number
   /**
    *
    * @type {number}
    * @memberof PaginatedResponseItemSimpleDto
    */
   totalPage?: number
   /**
    *
    * @type {Array<ItemSimpleDto>}
    * @memberof PaginatedResponseItemSimpleDto
    */
   items?: Array<ItemSimpleDto> | null
}